Do not generate all the programs with a prefix outcome for which we know.
For example if the program starts with a `-`, we know that it will immediately overflow. Before this change we still went through all the programs starting with a minus, tried to run each one and only then noticed that it is overflowing.

Using this speedup, we ran all programs of length 14.