Start looking for proofs at some places in the middle of the program. Also update the results up to 14.
For example, if we see a sequence `+]` in the middle of the program, then the loop that ends with this `]` will run forever, provided that we reach the `+` in the first place. So what we do, is we start constructing invariants from a place in the middle of the program. If we don't manage to prove that it doesn't halt, we drop this part of the proof. But if it does, we keep those invariants and look for another place earlier in the program to start the proof. Once we reach the place the program where we already have the invariants, either we skip this whole place, or we prove them.